pyproject.toml
[tool.poetry]
name = "cgmes2linkml"
version = "0.1.0"
description = ""
authors = ["Bart Kleijngeld <bartkl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
xmltodict = "^0.13.0"
pydantic = "^2.5.3"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
